user_service/requirements.txt
Flask
gunicorn
Flask-SQLAlchemy
psycopg2-binary
Werkzeug
PyJWT